fix(product-client): move workspace-switch shortcuts behind a lazy chunk to restore login first-load headroom
Main is currently at a knife edge on the "Login first-load budget (phase-6)"
CI gate: jsBytes was 500263 B against a cap of 502000 B, only ~1.7KB of local
headroom (less once CI's ~0.36% gzip overestimate vs local is applied). Every
open frontend PR was coin-flipping the gate by +/-10 bytes of minifier noise
(receipts: PR #2005 measured 502004, +4 over; #1992 measured 502008, +8 over;

Root cause: useAppShortcuts, mounted unconditionally by ProductLifecycleRoot,
called useWorkspaceNavigationWorkflow() to back three shortcuts
(workspace.by-index, workspace.previous-workspace, workspace.next-workspace)
and their held-key traversal cursor. That hook's selectWorkspaceFromSurface
pulls in workspace selection, the bundled agent-catalog JSON, and the
session-creation graph -- none of which can do anything before a workspace
exists, i.e. before auth -- so it all rode into the /login entry chunk anyway.

Fix: split those three handlers and the traversal-cursor machinery out of
useAppShortcuts into a new hook, useWorkspaceSwitchShortcuts, and mount it
only once authenticated via a new lazy component
(AuthenticatedWorkspaceSwitchShortcuts), the same authenticated-only + lazy
treatment ProductLifecycleRoot already gives AuthRestartOfferRoot,
SupportReportQueueRoot, AuthenticatedLaunchLifecycles, and
AuthenticatedBackgroundLifecycles. The shortcuts were already no-ops signed
out (no sidebar targets, no committed workspace to step from), so this changes
only when the handlers are registered, never what they do.

Local measurement: index chunk gzip JS 500263 B -> 495028 B (a 5235 B / ~5.1KB
reduction), moving headroom against the 502000 B cap from 1737 B to 6972 B.
Existing use-app-shortcuts tests for the moved shortcuts were relocated
unmodified in assertions to a new use-workspace-switch-shortcuts.test.tsx;
ProductLifecycleRoot.test.tsx gets one added mock for the new lazy component,
matching its existing AuthRestartOfferRoot/AuthenticatedBackgroundLifecycles
mocks.
